iotconnect: remove trailing whitespaces
Remove trailing whitespaces from SDK's code. Follow-up commits will come
addressing this issue in the other files. This avoids having a single
huge commit and easens code review.

Trailing whitespaces lead to "false positives" in git diff and can be
quite annoying when editing the code, not to mention they are supposed
to be avoided according to pep8.

Signed-off-by: Joao Marcos Costa <jcosta@witekio.com>